+++ libraries/ui_compose_material3/src/main/java/androidx/media3/ui/compose/material3/PlayerTokens.kt
@@ -19,6 +19,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 
@@ -30,6 +31,11 @@ internal object PlayerTokens {
   val CenterControlsSpacing = 10.dp
   val ControlsHorizontalPadding = 15.dp
   val BottomControlsHorizontalPadding = 8.dp
+  val ErrorTextStyle: TextStyle
+    @Composable get() = MaterialTheme.typography.titleMedium
+
+  // Approx MediumButtonSize/2 + 30.dp + ErrorTextStyle.lineHeight/2
+  val ErrorVerticalOffsetFromControls: Dp = (-80).dp
 
   val bottomControlsGradient: Brush
     @Composable